+++ girder/girder_large_image/web_client/main.js
@@ -2,6 +2,7 @@ import { registerPluginNamespace } from '@girder/core/pluginUtils';
 
 // import modules for side effects
 import './routes';
+import './eventStream';
 import './views/fileList';
 import './views/itemList';
 import './views/itemView';