+++ girder/test_girder/test_web_client.py
@@ -30,3 +30,19 @@ def testWebClient(boundServer, fsAssetstore, db, spec, girderWorker):
 def testWebClientNoWorker(boundServer, fsAssetstore, db, spec):
     spec = os.path.join(os.path.dirname(__file__), 'web_client_specs', spec)
     runWebClientTest(boundServer, spec, 15000)
+
+
+@pytest.mark.singular
+@pytest.mark.usefixtures('unbindLargeImage')
+@pytest.mark.plugin('large_image')
+@pytest.mark.parametrize('spec', (
+    'imageViewerSpec.js',
+))
+def testWebClientNoStream(boundServer, fsAssetstore, db, spec, girderWorker):
+    from girder.models.setting import Setting
+    from girder.settings import SettingKey
+
+    Setting().set(SettingKey.ENABLE_NOTIFICATION_STREAM, False)
+
+    spec = os.path.join(os.path.dirname(__file__), 'web_client_specs', spec)
+    runWebClientTest(boundServer, spec, 15000)